# docs/swagger.yaml
basePath: /
definitions:
  auth.CreateUserInput:
    properties:
      email:
        type: string
      fullname:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - fullname
    - password
    - phone
    type: object
  auth.ResendEmailOtpInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  database.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      fullname:
        type: string
      id:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  util.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  util.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  util.SuccessResponseWithData-database_User:
    properties:
      data:
        $ref: '#/definitions/database.User'
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Blanq Invoice API
  version: "1.0"
paths:
  /auth/resendEmailOtp:
    post:
      consumes:
      - application/json
      description: Resend EmailOTP
      parameters:
      - description: ' '
        in: body
        name: ResendEmailOtpInput
        required: true
        schema:
          $ref: '#/definitions/auth.ResendEmailOtpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: Resend EmailOTP
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: With the Bearer prefix
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: body
        name: CreateUserInput
        required: true
        schema:
          $ref: '#/definitions/auth.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.SuccessResponseWithData-database_User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: Signup
      tags:
      - Authentication
swagger: "2.0"